Analysis of police records for predictive policing
Mathematical approaches have  become more and more relevant in policing. In the CBS TV series, Numb3rs a mathematician genius helps his FBI agent brother to solve crime using maths. It turns out that much of the maths used in the series is actually relevant for law enforcement . There are many approaches described in the book "The numbers behind Numb3rs" and in the series : One of the most prominent method is geoprofiling.

It turns out that Los Angeles mathematicians work together with the police on a methods called predictive policing, where data and mathematical approaches are being used to direct police officers to locations where there is a high probability of a crime being committed. There has been a substantial drop in crime rates . The mathematician Prof. Andrea Bertozzi is leading the efforts in predictive policing.  There is a video where they discuss a model based on the (near) repeat victimisation theory - which is based on the observation that often the same or similar targets have a higher likelihood of being targeted again. In this project i try to analyse the data from Merseyside police  in order to locate the places in Liverpool where crime rates are high and the types of crimes being committed in those areas.